--- commands/commands.go
@@ -1,5 +1,0 @@
-// The individual commands for sweet
-//
-// main - Called with no subcommand
-// help - Opens a usage message
-package commands
